+++ src/lib/Components/Consignments/Screens/__tests__/SelectFromPhotoLibrary-tests.tsx
@@ -9,6 +9,8 @@ it("Sets up the right view hierarchy", () => {
   const nav = {} as any
   const route = {} as any
 
-  const tree = renderer.create(<SelectFromPhotoLibrary navigator={nav} route={route} />).toJSON()
+  const tree = renderer
+    .create(<SelectFromPhotoLibrary navigator={nav} route={route} setup={{}} updateWithPhotos={() => ""} />)
+    .toJSON()
   expect(tree).toMatchSnapshot()
 })