+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 -   Publish a consignment to convection - orta
 -   Keep track of the user's input during the submission process - orta
 -   Consolidate a lot of user interface elements - orta
+-   Photos are uploaded an attached to existing submissions  - orta
 
 ###### Messaging